refactor: split preprocess into files
