+++ packages/preprocess/package.json
@@ -13,6 +13,6 @@
     "unified": "^9.1.0",
     "unist-util-visit": "^2.0.3",
     "svelte": "3.55.0",
-    "svelte-preprocess": "^5.0.3"
+    "svelte-preprocess": "5.0.3"
   }
 }
